Restructure external apis
Update tests to use Environment variables. Add Folder structure to README
