+++ Organizer/__init__.py
@@ -6,41 +6,18 @@ __all__ = [
     "SeriesInfo",
     "load_info",
     "save_info",
-    "add_comicvine_info",
-    "add_league_info",
-    "add_metron_info",
     "Console",
-    "COLLECTION_FOLDER",
-    "COMICVINE_API_KEY",
-    "LEAGUE_API_KEY",
-    "LEAGUE_CLIENT_ID",
-    "METRON_PASSWORD",
-    "METRON_USERNAME",
-    "PROCESSING_FOLDER",
-    "del_folder",
-    "list_files",
+    "add_info",
+    "Settings",
     "slugify_comic",
     "slugify_publisher",
     "slugify_series",
 ]
 
-from .comic_archive import create_archive, extract_archive
-from .comic_info import ComicInfo, PublisherInfo, SeriesInfo, load_info, save_info
-from .comicvine_api import add_info as add_comicvine_info
-from .console import Console
-from .league_of_comic_geeks_api import add_info as add_league_info
-from .metron_api import add_info as add_metron_info
-from .utils import (
-    COLLECTION_FOLDER,
-    COMICVINE_API_KEY,
-    LEAGUE_API_KEY,
-    LEAGUE_CLIENT_ID,
-    METRON_PASSWORD,
-    METRON_USERNAME,
-    PROCESSING_FOLDER,
-    del_folder,
-    list_files,
-    slugify_comic,
-    slugify_publisher,
-    slugify_series,
-)
+from Organizer.comic_archive import create_archive, extract_archive
+from Organizer.comic_info import ComicInfo, PublisherInfo, SeriesInfo, load_info, save_info
+from Organizer.console import Console
+from Organizer.utils import slugify_comic, slugify_publisher, slugify_series
+
+from Organizer.settings import Settings  # isort:skip
+from Organizer.external import add_info  # isort:skip
