+++ Organizer/external/comicvine_api/__init__.py
@@ -1,0 +1,9 @@
+__all__ = ["add_info", "parse_issue_result", "parse_publisher_result", "parse_volume_result", "Talker"]
+
+from Organizer.external.comicvine_api.comicvine_talker import Talker  # isort:skip
+from Organizer.external.comicvine_api.comicvine_info import (
+    add_info,
+    parse_issue_result,
+    parse_publisher_result,
+    parse_volume_result,
+)
