+++ Organizer/external/league_of_comic_geeks_api/__init__.py
@@ -1,0 +1,4 @@
+__all__ = ["add_info", "parse_comic_result", "Talker"]
+
+from Organizer.external.league_of_comic_geeks_api.league_talker import Talker  # isort:skip
+from Organizer.external.league_of_comic_geeks_api.league_info import add_info, parse_comic_result
