--- Organizer/league_of_comic_geeks_api/__init__.py
@@ -1,3 +1,0 @@
-__all__ = ["Talker", "add_info", "parse_comic_result"]
-
-from Organizer.league_of_comic_geeks_api.league_info import Talker, add_info, parse_comic_result
