--- config-example.ini
@@ -1,13 +1,0 @@
-[General]
-Root =
-
-[Comicvine]
-API Key =
-
-[League of Comic Geeks]
-API Key =
-Client ID =
-
-[Metron]
-Username =
-Password =
