+++ tests/test_comicvine_api.py
@@ -1,7 +1,7 @@
 from datetime import date
 
-from Organizer.comic_info import ComicInfo, PublisherInfo, SeriesInfo
-from Organizer.comicvine_api import parse_issue_result, parse_publisher_result, parse_volume_result
+from Organizer import ComicInfo, PublisherInfo, SeriesInfo
+from Organizer.external.comicvine_api import parse_issue_result, parse_publisher_result, parse_volume_result
 
 PUBLISHER_ID = 10
 PUBLISHER_TITLE = "DC Comics"
