+++ tox.ini
@@ -3,6 +3,12 @@ isolated_build = true
 envlist = py38, py39
 
 [testenv]
+setenv =
+	COMICVINE_API_KEY={env:COMICVINE_API_KEY}
+	LEAGUE_API_KEY={env:LEAGUE_API_KEY}
+	LEAGUE_CLIENT_ID={env:LEAGUE_CLIENT_ID}
+	METRON_USERNAME={env:METRON_USERNAME}
+	METRON_PASSWORD={env:METRON_PASSWORD}
 deps = pytest
 commands =
 	python --version
